Refactor project source code structure.
