+++ .eslintcache
@@ -1,1 +1,1 @@
-[{"D:\\Hacker\\Projects\\karmayoga-web\\src\\reportWebVitals.js":"1","D:\\Hacker\\Projects\\karmayoga-web\\src\\App.js":"2","D:\\Hacker\\Projects\\karmayoga-web\\src\\index.js":"3","D:\\Hacker\\Projects\\karmayoga-web\\src\\BottomNavbar.js":"4","D:\\Hacker\\Projects\\karmayoga-web\\src\\Home.js":"5","D:\\Hacker\\Projects\\karmayoga-web\\src\\Calendar.js":"6","D:\\Hacker\\Projects\\karmayoga-web\\src\\Tasks.js":"7"},{"size":362,"mtime":1609034820726,"results":"8","hashOfConfig":"9"},{"size":676,"mtime":1609049313858,"results":"10","hashOfConfig":"9"},{"size":500,"mtime":1609034820726,"results":"11","hashOfConfig":"9"},{"size":1288,"mtime":1609048900046,"results":"12","hashOfConfig":"9"},{"size":456,"mtime":1609047739585,"results":"13","hashOfConfig":"9"},{"size":216,"mtime":1609050929209,"results":"14","hashOfConfig":"9"},{"size":140,"mtime":1609050953161,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"f623vr",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Hacker\\Projects\\karmayoga-web\\src\\reportWebVitals.js",[],["31","32"],"D:\\Hacker\\Projects\\karmayoga-web\\src\\App.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\index.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\BottomNavbar.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\Home.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\Calendar.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\Tasks.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
+[{"D:\\Hacker\\Projects\\karmayoga-web\\src\\reportWebVitals.js":"1","D:\\Hacker\\Projects\\karmayoga-web\\src\\App.js":"2","D:\\Hacker\\Projects\\karmayoga-web\\src\\index.js":"3","D:\\Hacker\\Projects\\karmayoga-web\\src\\components\\BottomNavbar.js":"4","D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Home.js":"5","D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Tasks.js":"6","D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Calendar.js":"7"},{"size":362,"mtime":1609034820726,"results":"8","hashOfConfig":"9"},{"size":705,"mtime":1617642872086,"results":"10","hashOfConfig":"9"},{"size":500,"mtime":1609034820726,"results":"11","hashOfConfig":"9"},{"size":1365,"mtime":1617642692096,"results":"12","hashOfConfig":"9"},{"size":456,"mtime":1617642625315,"results":"13","hashOfConfig":"9"},{"size":140,"mtime":1617642625462,"results":"14","hashOfConfig":"9"},{"size":216,"mtime":1617642625407,"results":"15","hashOfConfig":"9"},{"filePath":"16","messages":"17","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},"f623vr",{"filePath":"19","messages":"20","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"21","messages":"22","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0,"usedDeprecatedRules":"18"},{"filePath":"23","messages":"24","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"25","messages":"26","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"27","messages":"28","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},{"filePath":"29","messages":"30","errorCount":0,"warningCount":0,"fixableErrorCount":0,"fixableWarningCount":0},"D:\\Hacker\\Projects\\karmayoga-web\\src\\reportWebVitals.js",[],["31","32"],"D:\\Hacker\\Projects\\karmayoga-web\\src\\App.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\index.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\components\\BottomNavbar.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Home.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Tasks.js",[],"D:\\Hacker\\Projects\\karmayoga-web\\src\\pages\\Calendar.js",[],{"ruleId":"33","replacedBy":"34"},{"ruleId":"35","replacedBy":"36"},"no-native-reassign",["37"],"no-negated-in-lhs",["38"],"no-global-assign","no-unsafe-negation"]
