+++ package.json
@@ -13,6 +13,7 @@
     "react-router": "^5.2.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "4.0.1",
+    "tui-calendar": "^1.13.0",
     "web-vitals": "^0.2.4"
   },
   "scripts": {
