+++ src/__tests__/App.test.js
@@ -1,5 +1,5 @@
 import { render } from '@testing-library/react';
-import App from './App';
+import App from '../App';
 
 test('renders without crashing', () => {
   render(<App />);
